Refactor obejct endpoint tests

+++ test/acceptance_with_python/rbac/test_object_endpoint.py
@@ -3,65 +3,40 @@ import weaviate
 import weaviate.classes as wvc
 from weaviate.rbac.models import RBAC
 from _pytest.fixtures import SubRequest
-from .conftest import _sanitize_role_name
+from .conftest import _sanitize_role_name, Role_Wrapper_Type, generate_missing_permissions
 
 pytestmark = pytest.mark.xdist_group(name="rbac")
 
 
-def test_obj_insert(request: SubRequest, admin_client, custom_client):
+def test_obj_insert(
+    request: SubRequest, admin_client, custom_client, role_wrapper: Role_Wrapper_Type
+):
     name = _sanitize_role_name(request.node.name)
     admin_client.collections.delete(name)
     admin_client.roles.delete(name)
     col = admin_client.collections.create(name=name)
-
-    admin_client.roles.create(
-        name=name,
-        permissions=[
-            RBAC.permissions.data.create(collection=col.name),
-            RBAC.permissions.config.read(collection=col.name),
-        ],
-    )
-    admin_client.roles.assign(user="custom-user", roles=name)
-
-    source_no_rights = custom_client.collections.get(name)  # no network call => no RBAC check
-
-    source_no_rights.data.insert({})
-
-    admin_client.roles.revoke(user="custom-user", roles=name)
-    admin_client.roles.delete(name)
-
-    # no metadata read
-    role = admin_client.roles.create(
-        name=name,
-        permissions=RBAC.permissions.data.create(collection=col.name),
-    )
-    admin_client.roles.assign(user="custom-user", roles=role.name)
-
-    source_no_rights = custom_client.collections.get(name)  # no network call => no RBAC check
-    with pytest.raises(weaviate.exceptions.UnexpectedStatusCodeException) as e:
+    required_permissions = [
+        RBAC.permissions.data.create(collection=col.name),
+        RBAC.permissions.config.read(collection=col.name),
+    ]
+    with role_wrapper(admin_client, request, required_permissions):
+        source_no_rights = custom_client.collections.get(name)  # no network call => no RBAC check
         source_no_rights.data.insert({})
-    assert e.value.status_code == 403
 
-    admin_client.roles.revoke(user="custom-user", roles=role.name)
-    admin_client.roles.delete(role.name)
-
-    # no data create
-    role = admin_client.roles.create(
-        name=name,
-        permissions=RBAC.permissions.config.read(collection=col.name),
-    )
-    admin_client.roles.assign(user="custom-user", roles=role.name)
-
-    source_no_rights = custom_client.collections.get(name)  # no network call => no RBAC check
-    with pytest.raises(weaviate.exceptions.UnexpectedStatusCodeException) as e:
-        source_no_rights.data.insert({})
-    assert e.value.status_code == 403
-
-    admin_client.roles.revoke(user="custom-user", roles=role.name)
-    admin_client.roles.delete(role.name)
+    for permission in generate_missing_permissions(required_permissions):
+        with role_wrapper(admin_client, request, permission):
+            source_no_rights = custom_client.collections.get(
+                name
+            )  # no network call => no RBAC check
+            with pytest.raises(weaviate.exceptions.UnexpectedStatusCodeException) as e:
+                source_no_rights.data.insert({})
+            assert e.value.status_code == 403
+    admin_client.collections.delete(name)
 
 
-def test_obj_replace(request: SubRequest, admin_client, custom_client):
+def test_obj_replace(
+    request: SubRequest, admin_client, custom_client, role_wrapper: Role_Wrapper_Type
+):
     name = _sanitize_role_name(request.node.name)
     admin_client.collections.delete(name)
     admin_client.roles.delete(name)
@@ -69,54 +44,28 @@ def test_obj_replace(request: SubRequest, admin_client, custom_client):
 
     uuid_to_replace = col.data.insert({})
 
-    admin_client.roles.create(
-        name=name,
-        permissions=[
-            RBAC.permissions.data.update(collection=col.name),
-            RBAC.permissions.config.read(collection=col.name),
-        ],
-    )
-    admin_client.roles.assign(user="custom-user", roles=name)
-
-    source_no_rights = custom_client.collections.get(name)  # no network call => no RBAC check
-
-    source_no_rights.data.replace(uuid=uuid_to_replace, properties={})
-
-    admin_client.roles.revoke(user="custom-user", roles=name)
-    admin_client.roles.delete(name)
-
-    # no metadata read
-    admin_client.roles.create(
-        name=name,
-        permissions=RBAC.permissions.data.update(collection=col.name),
-    )
-    admin_client.roles.assign(user="custom-user", roles=name)
-
-    source_no_rights = custom_client.collections.get(name)  # no network call => no RBAC check
-    with pytest.raises(weaviate.exceptions.UnexpectedStatusCodeException) as e:
+    required_permissions = [
+        RBAC.permissions.data.update(collection=col.name),
+        RBAC.permissions.config.read(collection=col.name),
+    ]
+    with role_wrapper(admin_client, request, required_permissions):
+        source_no_rights = custom_client.collections.get(name)  # no network call => no RBAC check
         source_no_rights.data.replace(uuid=uuid_to_replace, properties={})
-    assert e.value.status_code == 403
 
-    admin_client.roles.revoke(user="custom-user", roles=name)
-    admin_client.roles.delete(name)
-
-    # no data create
-    admin_client.roles.create(
-        name=name, permissions=RBAC.permissions.config.read(collection=col.name)
-    )
-    admin_client.roles.assign(user="custom-user", roles=name)
-
-    source_no_rights = custom_client.collections.get(name)  # no network call => no RBAC check
-    with pytest.raises(weaviate.exceptions.UnexpectedStatusCodeException) as e:
-        source_no_rights.data.replace(uuid=uuid_to_replace, properties={})
-    assert e.value.status_code == 403
-
-    admin_client.roles.revoke(user="custom-user", roles=name)
-    admin_client.roles.delete(name)
+    for permission in generate_missing_permissions(required_permissions):
+        with role_wrapper(admin_client, request, permission):
+            source_no_rights = custom_client.collections.get(
+                name
+            )  # no network call => no RBAC check
+            with pytest.raises(weaviate.exceptions.UnexpectedStatusCodeException) as e:
+                source_no_rights.data.replace(uuid=uuid_to_replace, properties={})
+            assert e.value.status_code == 403
     admin_client.collections.delete(name)
 
 
-def test_obj_update(request: SubRequest, admin_client, custom_client):
+def test_obj_update(
+    request: SubRequest, admin_client, custom_client, role_wrapper: Role_Wrapper_Type
+):
     name = _sanitize_role_name(request.node.name)
     admin_client.collections.delete(name)
     admin_client.roles.delete(name)
@@ -124,55 +73,28 @@ def test_obj_update(request: SubRequest, admin_client, custom_client):
 
     uuid_to_replace = col.data.insert({})
 
-    admin_client.roles.create(
-        name=name,
-        permissions=[
-            RBAC.permissions.data.update(collection=col.name),
-            RBAC.permissions.config.read(collection=col.name),
-        ],
-    )
-    admin_client.roles.assign(user="custom-user", roles=name)
-
-    source_no_rights = custom_client.collections.get(name)  # no network call => no RBAC check
-
-    source_no_rights.data.update(uuid=uuid_to_replace, properties={})
-
-    admin_client.roles.revoke(user="custom-user", roles=name)
-    admin_client.roles.delete(name)
-
-    # no metadata read
-    admin_client.roles.create(
-        name=name,
-        permissions=RBAC.permissions.data.update(collection=col.name),
-    )
-    admin_client.roles.assign(user="custom-user", roles=name)
-
-    source_no_rights = custom_client.collections.get(name)  # no network call => no RBAC check
-    with pytest.raises(weaviate.exceptions.UnexpectedStatusCodeException) as e:
+    required_permissions = [
+        RBAC.permissions.data.update(collection=col.name),
+        RBAC.permissions.config.read(collection=col.name),
+    ]
+    with role_wrapper(admin_client, request, required_permissions):
+        source_no_rights = custom_client.collections.get(name)  # no network call => no RBAC check
         source_no_rights.data.update(uuid=uuid_to_replace, properties={})
-    assert e.value.status_code == 403
 
-    admin_client.roles.revoke(user="custom-user", roles=name)
-    admin_client.roles.delete(name)
-
-    # no data update
-    admin_client.roles.create(
-        name=name,
-        permissions=RBAC.permissions.config.read(collection=col.name),
-    )
-    admin_client.roles.assign(user="custom-user", roles=name)
-
-    source_no_rights = custom_client.collections.get(name)  # no network call => no RBAC check
-    with pytest.raises(weaviate.exceptions.UnexpectedStatusCodeException) as e:
-        source_no_rights.data.replace(uuid=uuid_to_replace, properties={})
-    assert e.value.status_code == 403
-
-    admin_client.roles.revoke(user="custom-user", roles=name)
-    admin_client.roles.delete(name)
+    for permission in generate_missing_permissions(required_permissions):
+        with role_wrapper(admin_client, request, permission):
+            source_no_rights = custom_client.collections.get(
+                name
+            )  # no network call => no RBAC check
+            with pytest.raises(weaviate.exceptions.UnexpectedStatusCodeException) as e:
+                source_no_rights.data.update(uuid=uuid_to_replace, properties={})
+            assert e.value.status_code == 403
     admin_client.collections.delete(name)
 
 
-def test_obj_delete(request: SubRequest, admin_client, custom_client):
+def test_obj_delete(
+    request: SubRequest, admin_client, custom_client, role_wrapper: Role_Wrapper_Type
+):
     name = _sanitize_role_name(request.node.name)
     admin_client.collections.delete(name)
     admin_client.roles.delete(name)
@@ -180,61 +102,32 @@ def test_obj_delete(request: SubRequest, admin_client, custom_client):
 
     uuid_to_delete = col.data.insert({})
 
-    admin_client.roles.create(
-        name=name,
-        permissions=[
-            RBAC.permissions.data.delete(collection=col.name),
-            RBAC.permissions.config.read(collection=col.name),
-        ],
-    )
-    admin_client.roles.assign(user="custom-user", roles=name)
+    required_permissions = [
+        RBAC.permissions.data.delete(collection=col.name),
+        RBAC.permissions.config.read(collection=col.name),
+    ]
+    with role_wrapper(admin_client, request, required_permissions):
+        col_no_rights = custom_client.collections.get(name)  # no network call => no RBAC check
+        assert len(col) == 1
+        col_no_rights.data.delete_by_id(uuid=uuid_to_delete)
+        assert len(col) == 0
 
-    col_no_rights = custom_client.collections.get(name)  # no network call => no RBAC check
-
-    assert len(col) == 1
-    col_no_rights.data.delete_by_id(uuid=uuid_to_delete)
-    assert len(col) == 0
-    admin_client.roles.revoke(user="custom-user", roles=name)
-    admin_client.roles.delete(name)
-
-    # no metadata read
     uuid_to_delete = col.data.insert({})
+    for permission in generate_missing_permissions(required_permissions):
+        with role_wrapper(admin_client, request, permission):
+            col_no_rights = custom_client.collections.get(name)  # no network call => no RBAC check
 
-    admin_client.roles.create(
-        name=name,
-        permissions=RBAC.permissions.data.delete(collection=col.name),
-    )
-    admin_client.roles.assign(user="custom-user", roles=name)
-
-    col_no_rights = custom_client.collections.get(name)  # no network call => no RBAC check
-
-    assert len(col) == 1
-    with pytest.raises(weaviate.exceptions.UnexpectedStatusCodeException) as e:
-        col_no_rights.data.delete_by_id(uuid=uuid_to_delete)
-    assert e.value.status_code == 403
-    assert len(col) == 1
-
-    admin_client.roles.revoke(user="custom-user", roles=name)
-    admin_client.roles.delete(name)
-
-    # no data delete
-    admin_client.roles.create(
-        name=name,
-        permissions=RBAC.permissions.config.read(collection=col.name),
-    )
-    admin_client.roles.assign(user="custom-user", roles=name)
-
-    col_no_rights = custom_client.collections.get(name)  # no network call => no RBAC check
-    with pytest.raises(weaviate.exceptions.UnexpectedStatusCodeException) as e:
-        col_no_rights.data.delete_by_id(uuid=uuid_to_delete)
-    assert e.value.status_code == 403
-
-    admin_client.roles.revoke(user="custom-user", roles=name)
-    admin_client.roles.delete(name)
+            assert len(col) == 1
+            with pytest.raises(weaviate.exceptions.UnexpectedStatusCodeException) as e:
+                col_no_rights.data.delete_by_id(uuid=uuid_to_delete)
+            assert e.value.status_code == 403
+            assert len(col) == 1
     admin_client.collections.delete(name)
 
 
-def test_obj_exists(request: SubRequest, admin_client, custom_client):
+def test_obj_exists(
+    request: SubRequest, admin_client, custom_client, role_wrapper: Role_Wrapper_Type
+):
     name = _sanitize_role_name(request.node.name)
     admin_client.collections.delete(name)
     admin_client.roles.delete(name)
@@ -242,49 +135,22 @@ def test_obj_exists(request: SubRequest, admin_client, custom_client):
 
     uuid_to_check = col.data.insert({})
 
-    admin_client.roles.create(
-        name=name,
-        permissions=[
-            RBAC.permissions.data.read(collection=col.name),
-            RBAC.permissions.config.read(collection=col.name),
-        ],
-    )
-    admin_client.roles.assign(user="custom-user", roles=name)
-
-    col_no_rights = custom_client.collections.get(name)  # no network call => no RBAC check
-
-    assert col_no_rights.data.exists(uuid=uuid_to_check)
+    required_permissions = [
+        RBAC.permissions.data.read(collection=col.name),
+        RBAC.permissions.config.read(collection=col.name),
+    ]
+    with role_wrapper(admin_client, request, required_permissions):
+        col_no_rights = custom_client.collections.get(name)  # no network call => no RBAC check
+        assert col_no_rights.data.exists(uuid=uuid_to_check)
     admin_client.roles.revoke(user="custom-user", roles=name)
     admin_client.roles.delete(name)
 
-    # no metadata read
-    admin_client.roles.create(
-        name=name,
-        permissions=RBAC.permissions.data.read(collection=col.name),
-    )
-    admin_client.roles.assign(user="custom-user", roles=name)
+    for permission in generate_missing_permissions(required_permissions):
+        with role_wrapper(admin_client, request, permission):
 
-    col_no_rights = custom_client.collections.get(name)  # no network call => no RBAC check
+            col_no_rights = custom_client.collections.get(name)  # no network call => no RBAC check
 
-    with pytest.raises(weaviate.exceptions.UnexpectedStatusCodeException) as e:
-        col_no_rights.data.exists(uuid=uuid_to_check)
-    assert e.value.status_code == 403
-
-    admin_client.roles.revoke(user="custom-user", roles=name)
-    admin_client.roles.delete(name)
-
-    # no data read
-    both_write = admin_client.roles.create(
-        name=name,
-        permissions=RBAC.permissions.config.read(collection=col.name),
-    )
-    admin_client.roles.assign(user="custom-user", roles=both_write.name)
-
-    col_no_rights = custom_client.collections.get(name)  # no network call => no RBAC check
-    with pytest.raises(weaviate.exceptions.UnexpectedStatusCodeException) as e:
-        col_no_rights.data.exists(uuid=uuid_to_check)
-    assert e.value.status_code == 403
-
-    admin_client.roles.revoke(user="custom-user", roles=both_write.name)
-    admin_client.roles.delete(both_write.name)
+            with pytest.raises(weaviate.exceptions.UnexpectedStatusCodeException) as e:
+                col_no_rights.data.exists(uuid=uuid_to_check)
+            assert e.value.status_code == 403
     admin_client.collections.delete(name)
